requirements.txt
ragas
datasets
python-dotenv
jsonschema
langchain
langchain-community
chromadb
google-generativeai
python-dotenv